Split out observers::storage module
